requirements.txt
aiofiles==0.5.0
aiomysql==0.0.21
amqp==5.0.2
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
bcrypt==3.2.0
billiard==3.6.3.0
celery==5.0.3
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
dnspython==2.0.0
email-validator==1.1.2
fastapi==0.62.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.9.0
httptools==0.1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
orjson==3.4.5   # 这个有问题
passlib==1.7.4
peewee==3.14.0
peewee-migrations==0.3.19
promise==2.3
prompt-toolkit==3.0.8
pycparser==2.20
pydantic==1.7.3
PyJWT==1.7.1
PyMySQL==0.9.3
python-multipart==0.0.5
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
requests==2.25.0
Rx==1.6.1
six==1.15.0
starlette==0.13.6
ujson==3.2.0
urllib3==1.26.2
uvicorn==0.11.8
uvloop==0.14.0
vine==5.0.0
wcwidth==0.2.5
websockets==8.1
